move everything under app
