+++ app/controllers/auth/index.js
